Reworked file structure
