+++ src/lobby/README.md
@@ -1,0 +1,6 @@
+Lobby
+=====
+
+The lobby folder contains all the parts for a functional instance
+of a game. It can be extended by passing custom classes to the
+`Lobby` constructor.
