+++ src/utils/NetCode.js
@@ -1,7 +1,10 @@
+/**
+ * Class that contains methods to communicate with other parties
+ */
 class NetCode {
 
     /**
-     * Send command to a Player
+     * Send command to a Client (e.g.: Player)
      * @param {Command} command
      * @param {socket} from
      * @param {socket} to
